requirements.txt
cassandra-driver
fastapi
gunicorn
jupyter
pandas
pypyr
python-dotenv
tensorflow
uvicorn
scikit-learn
boto3